# specs/001-build-a-website/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Domaine Vallot E-Commerce API
  description: API for boutique winery direct-to-consumer e-commerce website
  version: 1.0.0
  contact:
    name: Domaine Vallot
    email: info@domainevallot.com
servers:
  - url: https://domainevallot.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /orders:
    get:
      summary: Get customer orders
      description: Retrieve order history for authenticated customer
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Order history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Create new order
      description: Create a new order from cart contents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shipping_address
                - billing_address
                - shipping_method
              properties:
                shipping_address:
                  $ref: '#/components/schemas/Address'
                billing_address:
                  $ref: '#/components/schemas/Address'
                shipping_method:
                  type: string
                notes:
                  type: string
                  maxLength: 500
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
                  payment_url:
                    type: string
                    format: uri
        '400':
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_number:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled, refunded]
        total_eur:
          type: number
          format: decimal
        created_at:
          type: string
          format: date-time
        shipping_address:
          $ref: '#/components/schemas/Address'
        tracking_number:
          type: string
...
    Address:
      type: object
      required:
        - first_name
        - last_name
        - address_line1
        - city
        - postal_code
        - country_code
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        company:
          type: string
          maxLength: 100
        address_line1:
          type: string
          maxLength: 100
        address_line2:
          type: string
          maxLength: 100
        city:
          type: string
          maxLength: 50
        state_province:
          type: string
          maxLength: 50
        postal_code:
          type: string
          maxLength: 20
        country_code:
          type: string
          pattern: '^[A-Z]{2}$'
        phone:
          type: string
          maxLength: 20
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object